Split security by provider as it was too large
