+++ src/security/jwks_bearer.rs
@@ -1,168 +1,6 @@
-//! # Security Module
-//!
-//! The security module provides authentication and authorization providers for BRRTRouter,
-//! implementing various security schemes defined in OpenAPI specifications.
-//!
-//! ## Overview
-//!
-//! This module implements the [`SecurityProvider`] trait for common authentication methods:
-//! - **API Keys** - Header, query parameter, or cookie-based API keys
-//! - **Bearer JWT** - JSON Web Token validation with signature verification
-//! - **OAuth2** - OAuth2 token validation with scope checking
-//!
-//! Security providers are registered with the application and automatically enforced based on
-//! the `security` requirements defined in your OpenAPI specification.
-//!
-//! ## Architecture
-//!
-//! Security validation follows this flow:
-//!
-//! 1. Request arrives with credentials (header, cookie, query param)
-//! 2. Router determines which security scheme(s) are required for the route
-//! 3. Appropriate [`SecurityProvider`] is invoked to validate credentials
-//! 4. If validation succeeds, request proceeds to handler
-//! 5. If validation fails, 401/403 response is returned
-//!
-//! ## Security Providers
-//!
-//! ### API Key Provider
-//!
-//! Validates simple API keys from headers, query parameters, or cookies:
-//!
-//! ```rust
-//! use brrtrouter::security::{SecurityProvider, SecurityRequest};
-//! use brrtrouter::spec::SecurityScheme;
-//!
-//! // Simple static API key validation
-//! struct ApiKeyProvider { key: String }
-//!
-//! impl SecurityProvider for ApiKeyProvider {
-//!     fn validate(&self, scheme: &SecurityScheme, scopes: &[String], req: &SecurityRequest) -> bool {
-//!         req.get_header("x-api-key")
-//!             .map(|k| k == self.key)
-//!             .unwrap_or(false)
-//!     }
-//! }
-//! ```
-//!
-//! ### Bearer JWT Provider
-//!
-//! The [`BearerJwtProvider`] validates JWTs with:
-//! - Signature verification
-//! - Scope checking
-//! - Cookie or header extraction
-//!
-//! ```rust
-//! use brrtrouter::security::BearerJwtProvider;
-//!
-//! let provider = BearerJwtProvider::new("my-secret-signature")
-//!     .cookie_name("auth_token");
-//! ```
-//!
-//! ### OAuth2 Provider
-//!
-//! The [`OAuth2Provider`] validates OAuth2 tokens with scope checking:
-//!
-//! ```rust
-//! use brrtrouter::security::OAuth2Provider;
-//!
-//! let provider = OAuth2Provider::new("oauth-signature");
-//! ```
-//!
-//! ## Caching
-//!
-//! Security providers support optional caching to reduce validation overhead:
-//! - Positive results can be cached to avoid repeated database/API lookups
-//! - Negative results can be cached to prevent brute force attacks
-//! - TTL-based expiration ensures credentials are re-validated periodically
-//!
-//! ## JWT Claims Extraction (BFF Pattern)
-//!
-//! For Backend-for-Frontend (BFF) architectures where a BFF service needs to forward
-//! user context to downstream microservices, BRRTRouter provides JWT claims extraction.
-//!
-//! When a JWT token is successfully validated, the decoded claims are automatically
-//! made available to handlers via `HandlerRequest::jwt_claims`. This enables:
-//!
-//! 1. **Accessing user information** in handlers (e.g., user ID, email, roles)
-//! 2. **Forwarding claims to downstream services** as headers or in request bodies
-//! 3. **Making authorization decisions** based on claims
-//! 4. **Logging user context** for observability
-//!
-//! ### Example: Accessing Claims in Handlers
-//!
-//! ```rust,no_run
-//! use brrtrouter::dispatcher::HandlerRequest;
-//!
-//! fn handler(req: HandlerRequest) {
-//!     if let Some(claims) = &req.jwt_claims {
-//!         let user_id = claims.get("sub").and_then(|v| v.as_str());
-//!         let email = claims.get("email").and_then(|v| v.as_str());
-//!         let org_id = claims.get("org_id").and_then(|v| v.as_str());
-//!
-//!         // Use claims for business logic
-//!         println!("User {} ({}) from org {}", 
-//!                  user_id.unwrap_or("unknown"),
-//!                  email.unwrap_or("unknown"),
-//!                  org_id.unwrap_or("unknown"));
-//!     }
-//! }
-//! ```
-//!
-//! ### Example: Forwarding Claims to Downstream Services
-//!
-//! ```rust,no_run
-//! use brrtrouter::dispatcher::HandlerRequest;
-//! use reqwest::blocking::Client;
-//!
-//! fn bff_handler(req: HandlerRequest) -> Result<(), Box<dyn std::error::Error>> {
-//!     let client = Client::new();
-//!     let mut downstream_req = client.get("http://downstream-service/api/data");
-//!
-//!     // Forward JWT token (Option 1: Forward original token)
-//!     if let Some(token) = req.get_header("authorization") {
-//!         downstream_req = downstream_req.header("Authorization", token);
-//!     }
-//!
-//!     // Forward claims as headers (Option 2: Extract and forward claims)
-//!     if let Some(claims) = &req.jwt_claims {
-//!         if let Some(user_id) = claims.get("sub").and_then(|v| v.as_str()) {
-//!             downstream_req = downstream_req.header("X-User-ID", user_id);
-//!         }
-//!         if let Some(email) = claims.get("email").and_then(|v| v.as_str()) {
-//!             downstream_req = downstream_req.header("X-User-Email", email);
-//!         }
-//!     }
-//!
-//!     let response = downstream_req.send()?;
-//!     // ... handle response
-//!     Ok(())
-//! }
-//! ```
-//!
-//! ### Claims Cache Performance
-//!
-//! JWT claims are cached after validation to avoid repeated decoding. The cache:
-//! - Uses LRU eviction when capacity is reached
-//! - Respects token expiration (with leeway)
-//! - Invalidates on key rotation (via `kid` in cache key)
-//! - Provides cache statistics via `JwksBearerProvider::cache_stats()`
-//!
-//! ## Example
-//!
-//! ```rust,ignore
-//! // Example: Register a security provider (requires full server setup)
-//! use brrtrouter::server::AppService;
-//! use brrtrouter::security::BearerJwtProvider;
-//! use std::sync::Arc;
-//!
-//! let jwt_provider = BearerJwtProvider::new("secret");
-//! service.register_security_provider("bearerAuth", Arc::new(jwt_provider));
-//! ```
-
-use crate::dispatcher::HeaderVec;
-use crate::router::ParamVec;
+use crate::security::{CacheStats, SecurityProvider, SecurityRequest};
 use crate::spec::SecurityScheme;
+use base64::Engine as _;
 use lru::LruCache;
 use std::collections::HashMap;
 use std::num::NonZeroUsize;
@@ -171,388 +9,6 @@ use std::sync::Arc;
 use std::time::{Duration, Instant};
 use tracing::{debug, warn};
 use url::Url;
-
-/// Cache statistics for JWT claims cache
-#[derive(Debug, Clone, Copy)]
-pub struct CacheStats {
-    /// Number of cache hits (successful lookups)
-    pub hits: u64,
-    /// Number of cache misses (lookups that required decode)
-    pub misses: u64,
-    /// Number of entries evicted due to LRU capacity
-    pub evictions: u64,
-    /// Current number of entries in cache
-    pub size: usize,
-    /// Maximum capacity of cache
-    pub capacity: usize,
-}
-
-impl CacheStats {
-    /// Calculate cache hit rate as a percentage
-    #[must_use]
-    pub fn hit_rate(&self) -> f64 {
-        let total = self.hits + self.misses;
-        if total == 0 {
-            0.0
-        } else {
-            (self.hits as f64 / total as f64) * 100.0
-        }
-    }
-}
-
-/// Request context for security validation.
-///
-/// Contains extracted credentials from various sources (headers, query, cookies)
-/// that security providers can use to validate the request.
-///
-/// # JSF Compliance
-///
-/// Uses SmallVec (HeaderVec/ParamVec) references to avoid copying request data.
-pub struct SecurityRequest<'a> {
-    /// HTTP headers from the request (SmallVec for stack allocation)
-    pub headers: &'a HeaderVec,
-    /// Query parameters from the request URL (SmallVec for stack allocation)
-    pub query: &'a ParamVec,
-    /// Cookies from the request (SmallVec for stack allocation)
-    pub cookies: &'a HeaderVec,
-}
-
-impl<'a> SecurityRequest<'a> {
-    /// Get a header by name (case-insensitive)
-    #[inline]
-    pub fn get_header(&self, name: &str) -> Option<&str> {
-        self.headers
-            .iter()
-            .find(|(k, _)| k.eq_ignore_ascii_case(name))
-            .map(|(_, v)| v.as_str())
-    }
-
-    /// Get a query parameter by name
-    #[inline]
-    pub fn get_query(&self, name: &str) -> Option<&str> {
-        self.query
-            .iter()
-            .find(|(k, _)| k.as_ref() == name)
-            .map(|(_, v)| v.as_str())
-    }
-
-    /// Get a cookie by name
-    #[inline]
-    pub fn get_cookie(&self, name: &str) -> Option<&str> {
-        self.cookies
-            .iter()
-            .find(|(k, _)| k.as_ref() == name)
-            .map(|(_, v)| v.as_str())
-    }
-}
-
-/// Trait for implementing security validation providers.
-///
-/// Implement this trait to create custom authentication/authorization logic
-/// for your OpenAPI security schemes.
-pub trait SecurityProvider: Send + Sync {
-    /// Validate a request against a security scheme.
-    ///
-    /// # Arguments
-    ///
-    /// * `scheme` - The OpenAPI security scheme definition
-    /// * `scopes` - Required scopes for this operation (for OAuth2/OpenID)
-    /// * `req` - The security request context with credentials
-    ///
-    /// # Returns
-    ///
-    /// `true` if the request is authenticated and authorized, `false` otherwise
-    fn validate(&self, scheme: &SecurityScheme, scopes: &[String], req: &SecurityRequest) -> bool;
-
-    /// Extract claims from a validated request (optional).
-    ///
-    /// This method is called after `validate()` returns `true` to extract any
-    /// claims or user information from the validated credentials. For JWT-based
-    /// providers, this returns the decoded JWT claims. For other providers, this
-    /// may return `None` or provider-specific information.
-    ///
-    /// # Arguments
-    ///
-    /// * `scheme` - The OpenAPI security scheme definition
-    /// * `req` - The security request context with credentials
-    ///
-    /// # Returns
-    ///
-    /// * `Some(Value)` - The extracted claims/information as JSON
-    /// * `None` - No claims available or provider doesn't support claims extraction
-    ///
-    /// # Default Implementation
-    ///
-    /// Returns `None` by default. Providers that support claims extraction should
-    /// override this method.
-    fn extract_claims(&self, scheme: &SecurityScheme, req: &SecurityRequest) -> Option<serde_json::Value> {
-        let _ = (scheme, req);
-        None
-    }
-}
-
-use base64::{engine::general_purpose, Engine as _};
-use serde_json::Value;
-
-/// Simple Bearer/JWT provider that validates tokens embedded in the
-/// `Authorization` header or a cookie.
-///
-/// Tokens are expected to have the form `header.payload.signature` where the
-/// signature part must match the configured `signature` string. Only the
-/// payload section is inspected for a whitespace separated `scope` field.
-pub struct BearerJwtProvider {
-    signature: String,
-    cookie_name: Option<String>,
-}
-
-impl BearerJwtProvider {
-    /// Create a new Bearer JWT provider with the given signature
-    ///
-    /// The signature is used to validate JWT tokens (checked against the 3rd part of the JWT).
-    /// This is a simplified implementation for testing - production should use proper JWT libraries.
-    ///
-    /// # Arguments
-    ///
-    /// * `signature` - Expected JWT signature value
-    pub fn new(signature: impl Into<String>) -> Self {
-        Self {
-            signature: signature.into(),
-            cookie_name: None,
-        }
-    }
-
-    /// Configure the cookie name used to read the token.
-    pub fn cookie_name(mut self, name: impl Into<String>) -> Self {
-        self.cookie_name = Some(name.into());
-        self
-    }
-
-    fn extract_token<'a>(&self, req: &'a SecurityRequest) -> Option<&'a str> {
-        if let Some(name) = &self.cookie_name {
-            if let Some(t) = req.get_cookie(name) {
-                return Some(t);
-            }
-        }
-        req.get_header("authorization")
-            .and_then(|h| h.strip_prefix("Bearer "))
-    }
-
-    fn validate_token(&self, token: &str, scopes: &[String]) -> bool {
-        let mut parts = token.split('.');
-        let header = parts.next();
-        let payload = parts.next();
-        let sig = parts.next();
-        if header.is_none() || payload.is_none() || sig != Some(self.signature.as_str()) {
-            debug!("BearerJWT token validation failed: malformed token or invalid signature");
-            return false;
-        }
-        // Safe to unwrap here because we checked is_none() above
-        let payload_bytes = match general_purpose::STANDARD
-            .decode(payload.expect("payload already validated as Some"))
-        {
-            Ok(b) => b,
-            Err(e) => {
-                debug!("BearerJWT token validation failed: invalid base64 payload - {:?}", e);
-                return false;
-            }
-        };
-        let json: Value = match serde_json::from_slice(&payload_bytes) {
-            Ok(v) => v,
-            Err(e) => {
-                debug!("BearerJWT token validation failed: invalid JSON payload - {:?}", e);
-                return false;
-            }
-        };
-        let token_scopes = json.get("scope").and_then(|v| v.as_str()).unwrap_or("");
-        let has_all_scopes = scopes
-            .iter()
-            .all(|s| token_scopes.split_whitespace().any(|ts| ts == s));
-        
-        if !has_all_scopes {
-            warn!(
-                "BearerJWT validation failed: missing required scopes (token: {:?}, required: {:?})",
-                token_scopes,
-                scopes
-            );
-        }
-        
-        has_all_scopes
-    }
-}
-
-/// Bearer JWT authentication provider implementation
-///
-/// Validates JWT tokens passed via the `Authorization: Bearer {token}` header.
-/// Checks token signature and required scopes.
-///
-/// # Validation Flow
-///
-/// 1. Verify security scheme is HTTP Bearer
-/// 2. Extract token from Authorization header (or cookie if configured)
-/// 3. Parse JWT and validate signature
-/// 4. Check if token contains all required scopes
-///
-/// # JWT Format
-///
-/// Expects JWT format: `header.payload.signature`
-/// - Signature (3rd part) must match configured signature value
-/// - Payload must be valid JSON with optional `scope` field
-/// - Scopes can be space-separated string: `"read:users write:users"`
-///
-/// # Security
-///
-/// This is a simplified implementation suitable for:
-/// - ✅ Testing and development
-/// - ✅ Internal microservices with pre-shared secrets
-/// - ❌ NOT for production with external clients (use `JwksBearerProvider`)
-///
-/// For production JWT validation with proper key rotation,
-/// use `JwksBearerProvider` with JWKS endpoint.
-impl SecurityProvider for BearerJwtProvider {
-    /// Validate a Bearer JWT token against the security scheme
-    ///
-    /// # Arguments
-    ///
-    /// * `scheme` - Security scheme from OpenAPI spec
-    /// * `scopes` - Required OAuth2 scopes from operation
-    /// * `req` - The security request containing headers/cookies
-    ///
-    /// # Returns
-    ///
-    /// - `true` - Token is valid and contains required scopes
-    /// - `false` - Token missing, invalid signature, or missing scopes
-    fn validate(&self, scheme: &SecurityScheme, scopes: &[String], req: &SecurityRequest) -> bool {
-        match scheme {
-            SecurityScheme::Http { scheme, .. } if scheme.eq_ignore_ascii_case("bearer") => {}
-            _ => {
-                debug!("BearerJWT validation failed: unsupported security scheme");
-                return false;
-            }
-        }
-        let token = match self.extract_token(req) {
-            Some(t) => t,
-            None => {
-                debug!("BearerJWT validation failed: missing token (no Authorization header or cookie)");
-                return false;
-            }
-        };
-        let result = self.validate_token(token, scopes);
-        if result {
-            debug!("BearerJWT validation succeeded: token valid");
-        }
-        result
-    }
-}
-
-/// OAuth2 provider using the same simple JWT validation as `BearerJwtProvider`.
-pub struct OAuth2Provider {
-    signature: String,
-    cookie_name: Option<String>,
-}
-
-impl OAuth2Provider {
-    /// Create a new OAuth2 provider with the given signature
-    ///
-    /// Uses JWT validation similar to `BearerJwtProvider`. This is a simplified
-    /// implementation for testing - production should use proper OAuth2 libraries.
-    ///
-    /// # Arguments
-    ///
-    /// * `signature` - Expected JWT signature value
-    pub fn new(signature: impl Into<String>) -> Self {
-        Self {
-            signature: signature.into(),
-            cookie_name: None,
-        }
-    }
-
-    /// Configure the cookie name used to read the OAuth2 token
-    ///
-    /// # Arguments
-    ///
-    /// * `name` - Cookie name (e.g., "oauth_token")
-    pub fn cookie_name(mut self, name: impl Into<String>) -> Self {
-        self.cookie_name = Some(name.into());
-        self
-    }
-
-    fn extract_token<'a>(&self, req: &'a SecurityRequest) -> Option<&'a str> {
-        if let Some(name) = &self.cookie_name {
-            if let Some(t) = req.get_cookie(name) {
-                return Some(t);
-            }
-        }
-        req.get_header("authorization")
-            .and_then(|h| h.strip_prefix("Bearer "))
-    }
-}
-
-/// OAuth2 provider implementation using JWT validation
-///
-/// Simplified OAuth2 provider that reuses `BearerJwtProvider` logic for token validation.
-/// Supports both Authorization header and cookie-based tokens.
-///
-/// # Validation Flow
-///
-/// 1. Verify security scheme is OAuth2
-/// 2. Extract token from cookie (if configured) or Authorization header
-/// 3. Delegate validation to `BearerJwtProvider` logic
-///
-/// # Token Sources (Priority Order)
-///
-/// 1. **Cookie**: If `cookie_name()` is configured, read token from cookie
-/// 2. **Authorization Header**: Falls back to `Authorization: Bearer {token}`
-///
-/// # Usage
-///
-/// ```rust
-/// use brrtrouter::security::OAuth2Provider;
-///
-/// // Authorization header only
-/// let provider = OAuth2Provider::new("secret_signature");
-///
-/// // Cookie-based (e.g., for browser SPAs)
-/// let provider = OAuth2Provider::new("secret_signature")
-///     .cookie_name("oauth_token");
-/// ```
-///
-/// # Security
-///
-/// - ✅ Testing and development
-/// - ✅ Internal APIs with controlled clients
-/// - ❌ NOT for production OAuth2 flows (use proper OAuth2 library)
-///
-/// For production: Use `JwksBearerProvider` with proper JWKS validation.
-impl SecurityProvider for OAuth2Provider {
-    /// Validate an OAuth2 token (uses JWT validation internally)
-    ///
-    /// # Arguments
-    ///
-    /// * `scheme` - Security scheme from OpenAPI spec (must be OAuth2)
-    /// * `scopes` - Required OAuth2 scopes from operation
-    /// * `req` - The security request containing headers/cookies
-    ///
-    /// # Returns
-    ///
-    /// - `true` - Token is valid and contains required scopes
-    /// - `false` - Token missing, invalid, or missing scopes
-    fn validate(&self, scheme: &SecurityScheme, scopes: &[String], req: &SecurityRequest) -> bool {
-        match scheme {
-            SecurityScheme::OAuth2 { .. } => {}
-            _ => return false,
-        }
-        let token = match self.extract_token(req) {
-            Some(t) => t,
-            None => return false,
-        };
-        // Reuse BearerJwtProvider logic
-        let helper = BearerJwtProvider {
-            signature: self.signature.clone(),
-            cookie_name: None,
-        };
-        helper.validate_token(token, scopes)
-    }
-}
 
 /// JWKS-based Bearer provider for production integrations.
 /// Fetches keys from a JWKS URL and validates JWTs (signature and claims).
@@ -1409,191 +865,5 @@ impl SecurityProvider for JwksBearerProvider {
             Ok(data) => Some(data.claims),
             Err(_) => None,
         }
-    }
-}
-
-/// Remote API key verification provider with simple caching.
-pub struct RemoteApiKeyProvider {
-    verify_url: String,
-    timeout_ms: u64,
-    cache_ttl: Duration,
-    cache: std::sync::Mutex<HashMap<String, (Instant, bool)>>,
-    header_name: String,
-}
-
-impl RemoteApiKeyProvider {
-    /// Create a new remote API key provider
-    ///
-    /// Validates API keys by making HTTP requests to an external verification service.
-    /// Results are cached to reduce latency and load on the verification service.
-    ///
-    /// # Arguments
-    ///
-    /// * `verify_url` - URL of the verification service (e.g., `https://api.example.com/verify`)
-    pub fn new(verify_url: impl Into<String>) -> Self {
-        Self {
-            verify_url: verify_url.into(),
-            timeout_ms: 500,
-            cache_ttl: Duration::from_secs(60),
-            cache: std::sync::Mutex::new(HashMap::new()),
-            header_name: "x-api-key".to_string(),
-        }
-    }
-
-    /// Configure the HTTP request timeout in milliseconds
-    ///
-    /// Default: 500ms
-    pub fn timeout_ms(mut self, ms: u64) -> Self {
-        self.timeout_ms = ms;
-        self
-    }
-
-    /// Configure the TTL for cached validation results
-    ///
-    /// Default: 60 seconds
-    pub fn cache_ttl(mut self, ttl: Duration) -> Self {
-        self.cache_ttl = ttl;
-        self
-    }
-
-    /// Configure the header name to look for the API key
-    ///
-    /// Default: `x-api-key`
-    pub fn header_name(mut self, name: impl Into<String>) -> Self {
-        self.header_name = name.into().to_ascii_lowercase();
-        self
-    }
-
-    fn extract_key<'a>(&self, req: &'a SecurityRequest, header_name: &str) -> Option<&'a str> {
-        // Prefer named header, also accept Authorization: Bearer <key>
-        req.get_header(header_name).or_else(|| {
-            req.get_header("authorization")
-                .and_then(|h| h.strip_prefix("Bearer "))
-        })
-    }
-}
-
-/// Remote API key provider implementation
-///
-/// Validates API keys by making HTTP requests to an external verification service.
-/// Implements caching to reduce latency and load on the verification endpoint.
-///
-/// # Validation Flow
-///
-/// 1. Verify security scheme is API Key with `location: header`
-/// 2. Extract API key from configured header (or Authorization: Bearer)
-/// 3. Check cache for recent validation result
-/// 4. If cache miss or expired: Make HTTP GET to verification URL
-/// 5. Cache result (success/failure) with TTL
-/// 6. Return validation result
-///
-/// # Key Extraction Priority
-///
-/// 1. **Custom header**: `X-API-Key` (or configured via `header_name()`)
-/// 2. **OpenAPI spec header**: Uses `name` from `securityScheme.name`
-/// 3. **Authorization header**: Falls back to `Authorization: Bearer {key}`
-///
-/// # Verification Request
-///
-/// Makes HTTP GET to configured URL with key in `X-API-Key` header:
-///
-/// ```text
-/// GET /verify HTTP/1.1
-/// Host: api.example.com
-/// X-API-Key: user_api_key_here
-/// ```
-///
-/// Success: 2xx status code
-/// Failure: Any other status code or timeout
-///
-/// # Caching
-///
-/// - **TTL**: Configurable (default: 60 seconds)
-/// - **Storage**: In-memory HashMap with Mutex
-/// - **Key**: API key string
-/// - **Value**: (timestamp, valid: bool)
-/// - **Eviction**: Lazy (checks on read)
-///
-/// # Usage
-///
-/// ```rust
-/// use brrtrouter::security::RemoteApiKeyProvider;
-/// use std::time::Duration;
-///
-/// let provider = RemoteApiKeyProvider::new("https://auth.example.com/verify")
-///     .timeout_ms(1000)                             // 1 second timeout
-///     .cache_ttl(Duration::from_secs(300))           // 5 minute cache
-///     .header_name("X-Custom-Key");                 // Custom header
-/// ```
-///
-/// # Performance
-///
-/// - **Cache hit**: ~1µs (HashMap lookup)
-/// - **Cache miss**: ~50-500ms (HTTP request)
-/// - Recommendation: Use longer TTL for trusted environments
-impl SecurityProvider for RemoteApiKeyProvider {
-    /// Validate an API key by making an HTTP request to verification service
-    ///
-    /// Uses caching to avoid repeated verification requests for the same key.
-    ///
-    /// # Arguments
-    ///
-    /// * `scheme` - Security scheme from OpenAPI spec (must be API Key)
-    /// * `_scopes` - Required scopes (unused - API keys don't have scopes)
-    /// * `req` - The security request containing headers
-    ///
-    /// # Returns
-    ///
-    /// - `true` - API key is valid (cached or verified remotely)
-    /// - `false` - API key missing, invalid, or verification failed
-    ///
-    /// # Cache Behavior
-    ///
-    /// - Cached success: Returns true immediately
-    /// - Cached failure: Returns false immediately
-    /// - Expired cache: Makes new verification request
-    /// - New key: Makes verification request and caches result
-    fn validate(&self, scheme: &SecurityScheme, _scopes: &[String], req: &SecurityRequest) -> bool {
-        let name = match scheme {
-            SecurityScheme::ApiKey { name, location, .. } if location == "header" => {
-                name.to_ascii_lowercase()
-            }
-            _ => return false,
-        };
-        let key = match self
-            .extract_key(req, &self.header_name)
-            .or_else(|| self.extract_key(req, &name))
-        {
-            Some(k) => k,
-            None => return false,
-        };
-        // Cache lookup
-        if let Some((ts, ok)) = self
-            .cache
-            .lock()
-            .expect("API key cache Mutex poisoned - critical error")
-            .get(key)
-            .cloned()
-        {
-            if ts.elapsed() < self.cache_ttl {
-                return ok;
-            }
-        }
-        // Remote verify
-        let client = reqwest::blocking::Client::builder()
-            .timeout(Duration::from_millis(self.timeout_ms))
-            .build();
-        let ok = match client {
-            Ok(c) => match c.get(&self.verify_url).header("X-API-Key", key).send() {
-                Ok(r) => r.status().is_success(),
-                Err(_) => false,
-            },
-            Err(_) => false,
-        };
-        self.cache
-            .lock()
-            .expect("API key cache Mutex poisoned - critical error")
-            .insert(key.to_string(), (Instant::now(), ok));
-        ok
     }
 }
